\docType{data}
\name{readings}
\alias{readings}
\title{readings}
\format{A list of length 10 - each item in the list is
the list of chapters for the given 'list'}
\usage{
readings
}
\description{
The chapters in each 'list' of Professor Grant Horners
Bible Reading System
}
\examples{
readings
}
\references{
\url{http://www.scribd.com/doc/12349985/Professor-Grant-Horners-Bible-Reading-System}
}

